Replace autogenerated properties for delay_on/off with custom ones
This introduces a workaround for ev3dev/ev3dev#225 and
fixes #234.

First, when trigger is set to 'timer', the setter blocks until
'delay_on' and 'delay_off' files have been created and got the
correct attributes. The maximum wait time is 1 second.

Second, since 'delay_on' and 'delay_off' attributes are created and
destroyed depending on value of 'trigger', their cache entries may
become outdated. This commit makes an attempt to reopen the files
in this case.